requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cloudinary==1.41.0
Django==5.1.4
djangorestframework==3.15.2
Faker==33.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
packaging==24.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.8.2
